src/bq_ch_migrator/cloud_function_template/requirements.txt
functions-framework>=3.0.0
google-cloud-bigquery>=3.41.0
